requirements.txt
fastapi
uvicorn
websockets
python-dotenv
web3
requests
gunicorn
pymongo
python-multipart
python-jose[cryptography]
passlib[bcrypt]